refactor(bootstrap): move folder inside src
